+++ src/de/omegazirkel/risingworld/tools/PluginSettings.java
@@ -25,7 +25,6 @@ public class PluginSettings {
 	public String logLevel = Level.DEBUG.name();
 	public boolean reloadOnChange = false;
 	public boolean enablePluginWelcomeMessage = false;
-	public boolean enableSleepAnnouncement = false;
 
 	// END Settings
 
@@ -74,12 +73,10 @@ public class PluginSettings {
 
 			// motd settings
 			enablePluginWelcomeMessage = settings.getProperty("enablePluginWelcomeMessage", "false").contentEquals("true");
-			enableSleepAnnouncement = settings.getProperty("enableSleepAnnouncement", "false").contentEquals("true");
 
 			logger().info(plugin.getName() + " Plugin settings loaded");
 			logger().info("Sending welcome message on login is: " + String.valueOf(enablePluginWelcomeMessage));
 			logger().info("enablePluginWelcomeMessage is: " + enablePluginWelcomeMessage);
-			logger().info("enableSleepAnnouncement is: " + enableSleepAnnouncement);
 			logger().info("Loglevel is set to " + logLevel);
 			logger().setLevel(logLevel);